Rename and move Arma Info
